requirements.txt
Flask==2.3.2
pandas==2.0.3
pg8000==1.29.8
python-dotenv==1.0.1
Requests==2.31.0
SQLAlchemy==2.0.23
Streamlit==1.29.0